- fixed some compiler warnings - video_x.cpp: now checks whether the X server has the XFree86DGA extension

+++ BasiliskII/src/macos_util.cpp
@@ -56,7 +56,7 @@ static bool is_drive_number_free(int num)
 	uint32 e = ReadMacInt32(0x308 + qHead);
 	while (e) {
 		uint32 d = e - dsQLink;
-		if (ReadMacInt16(d + dsQDrive) == num)
+		if ((int)ReadMacInt16(d + dsQDrive) == num)
 			return false;
 		e = ReadMacInt32(e + qLink);
 	}
